feat(git-ui): Add set remote url config

+++ git/git-ui/git-ui.js
@@ -10,25 +10,28 @@ const successInitMessage = 'init written successfully';
 module.exports = {
   commit: (userDir, message) => new Promise((resolve, reject) => {
     // deletes package.json
-    fs.unlink(`${userDir}/package.json`, (err) => {
-      if (err && err.code !== fileDoesNotExist) {
-        reject(err);
+
+    if (fs.existsSync(`${userDir}/package.json`)) {
+      fs.unlink(`${userDir}/package.json`, (err) => {
+        if (err && err.code !== fileDoesNotExist) {
+          reject(err);
+        }
+      });
+    }
+
+    // generates a new package.json based on node_modules
+    exec('npm init -y', { cwd: userDir }, (error, stdout, stderr) => {
+      if (error) {
+        reject(error);
+      } else if (stderr && !stderr.includes(successInitMessage)) {
+        reject(stderr);
       } else {
-        // generates a new package.json based on node_modules
-        exec('npm init -y', { cwd: userDir }, (error, stdout, stderr) => {
-          if (error) {
-            reject(error);
-          } else if (stderr && !stderr.includes(successInitMessage)) {
-            reject(stderr);
+        // commits and pushes all changes to the remote branch
+        git.add('--all').commit(message).push(remote, branch, (ex, data) => {
+          if (ex) {
+            reject(ex);
           } else {
-            // commits and pushes all changes to the remote branch
-            git.add('--all').commit(message).push(remote, branch, (ex, data) => {
-              if (ex) {
-                reject(ex);
-              } else {
-                resolve(data);
-              }
-            });
+            resolve(data);
           }
         });
       }
@@ -111,5 +114,25 @@ module.exports = {
     } catch (e) {
       reject(e);
     }
+  }),
+
+  remoteGet: () => new Promise((resolve, reject) => {
+    git.getRemotes(true, (err, result) => {
+      if (err) {
+        reject(err);
+      } else {
+        resolve(result[0].refs.push);
+      }
+    });
+  }),
+
+  remoteSet: url => new Promise((resolve, reject) => {
+    git.raw(['remote', 'set-url', remote, url], (err) => {
+      if (err) {
+        reject(err);
+      } else {
+        resolve(url);
+      }
+    });
   }),
 };
